Comm DeviceControllers are separated from CommChannelService file.
